requirements/test.txt
pytest>=5.4.1,<5.5
coverage>=5.1,<5.2
pytest-cov>=2.8.1,<2.9
mypy==0.770
requests>=2.23.0
vcrpy>=4.0.2
